@@ -4,13 +4,13 @@ import 'package:flutter/physics.dart';
 import 'package:flutter/widgets.dart';
 
 class ScaleTap extends StatefulWidget {
-  final Function() onTap;
-  final Function() onLongPress;
-  final Widget child;
+  final Function()? onTap;
+  final Function()? onLongPress;
+  final Widget? child;
   final Duration duration;
   final double scaleMinValue;
-  final Curve scaleCurve;
-  final Curve opacityCurve;
+  final Curve? scaleCurve;
+  final Curve? opacityCurve;
   final double opacityMinValue;
 
   ScaleTap({
@@ -28,10 +28,11 @@ class ScaleTap extends StatefulWidget {
   _ScaleTapState createState() => _ScaleTapState();
 }
 
-class _ScaleTapState extends State<ScaleTap> with SingleTickerProviderStateMixin {
-  AnimationController _animationController;
-  Animation<double> _scale;
-  Animation<double> _opacity;
+class _ScaleTapState extends State<ScaleTap>
+    with SingleTickerProviderStateMixin {
+  late final AnimationController _animationController;
+  late final Animation<double> _scale;
+  late final Animation<double> _opacity;
 
   @override
   void initState() {
@@ -52,7 +53,7 @@ class _ScaleTapState extends State<ScaleTap> with SingleTickerProviderStateMixin
 
   @override
   void dispose() {
-    _animationController?.dispose();
+    _animationController.dispose();
     super.dispose();
   }
 
@@ -65,12 +66,16 @@ class _ScaleTapState extends State<ScaleTap> with SingleTickerProviderStateMixin
   }
 
   Duration get _computedDuration {
-    return widget.duration ?? Duration(milliseconds: 300);
+    return widget.duration;
   }
 
-  Future<void> anim({double scale, double opacity, Duration duration}) {
-    _animationController?.stop();
-    _animationController.duration = duration ?? Duration.zero;
+  Future<void> anim({
+    required double scale,
+    required double opacity,
+    required Duration duration,
+  }) {
+    _animationController.stop();
+    _animationController.duration = duration;
 
     _scale = Tween<double>(
       begin: _scale.value,
@@ -86,8 +91,8 @@ class _ScaleTapState extends State<ScaleTap> with SingleTickerProviderStateMixin
       curve: _computedOpacityCurve,
       parent: _animationController,
     ));
-    _animationController?.reset();
-    return _animationController?.forward();
+    _animationController.reset();
+    return _animationController.forward();
   }
 
   Future<void> _onTapDown(_) {
@@ -110,7 +115,7 @@ class _ScaleTapState extends State<ScaleTap> with SingleTickerProviderStateMixin
     return _onTapUp(_);
   }
 
-  Widget _container({Widget child}) {
+  Widget _container({required Widget child}) {
     if (widget.onTap != null || widget.onLongPress != null) {
       return Listener(
         onPointerDown: _onTapDown,
@@ -132,7 +137,7 @@ class _ScaleTapState extends State<ScaleTap> with SingleTickerProviderStateMixin
     return _container(
       child: AnimatedBuilder(
         animation: _animationController,
-        builder: (_, Widget child) {
+        builder: (_, child) {
           return Opacity(
             opacity: _opacity.value,
             child: Transform.scale(
